+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 ## NEXT
 
+**BREAKING CHANGES:**
+
+* The initial state of the `MobileScannerState` camera facing direction is changed to `CameraFacing.unknown`.
+
 Improvements:
 * [Android] Turn off logging for CameraX, except for the `Log.ERROR` logging level.
 * Added `CameraFacing.external` and `CameraFacing.unknown` enum values.